scripts/service_alerts/requirements.txt
google-cloud-storage==1.33.0
oauth2client==4.1.3
